resolve signup flag server-side to avoid signup link flicker
