+++ ui/src/app/auth/login/page.tsx
@@ -1,113 +1,14 @@
-"use client";
+import { getSignupEnabled } from "@/lib/auth/config";
 
-import Link from "next/link";
-import { useEffect, useState } from "react";
-import { toast } from "sonner";
+import { LoginForm } from "./LoginForm";
 
-import { loginApiV1AuthLoginPost } from "@/client/sdk.gen";
-import { AuthEnterpriseCTA } from "@/components/auth/AuthEnterpriseCTA";
-import { AuthShell } from "@/components/auth/AuthShell";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
+// Resolve the backend health check before rendering so the "Sign up" link is
+// correct on first paint — no client-side fetch, no flicker on locked-down
+// installs. force-dynamic keeps the page off the build-time prerender, which
+// would bake in the flag's build-environment value.
+export const dynamic = "force-dynamic";
 
-export default function LoginPage() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [loading, setLoading] = useState(false);
-  // Default true (the backend default) so the link doesn't flash out on stock installs.
-  const [signupEnabled, setSignupEnabled] = useState<boolean>(true);
-
-  useEffect(() => {
-    let cancelled = false;
-    fetch("/api/config/auth")
-      .then((r) => (r.ok ? r.json() : null))
-      .then((data) => {
-        if (!cancelled) setSignupEnabled(data?.signupEnabled !== false);
-      })
-      .catch(() => {
-        // Backend unreachable — keep the default.
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-
-    try {
-      const res = await loginApiV1AuthLoginPost({
-        body: { email, password },
-      });
-
-      if (res.error || !res.data) {
-        const detail = (res.error as { detail?: string })?.detail;
-        toast.error(detail || "Login failed");
-        return;
-      }
-
-      // Set httpOnly cookies via server route
-      await fetch("/api/auth/session", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ token: res.data.token, user: res.data.user }),
-      });
-
-      window.location.href = "/after-sign-in";
-    } catch {
-      toast.error("An error occurred. Please try again.");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <AuthShell enterpriseSlot={<AuthEnterpriseCTA />}>
-      <div className="space-y-1.5 text-center">
-        <h1 className="text-2xl font-semibold tracking-tight">Sign in</h1>
-        <p className="text-sm text-muted-foreground">
-          Enter your email and password to continue
-        </p>
-      </div>
-
-      <form onSubmit={handleSubmit} className="space-y-4">
-        <div className="space-y-2">
-          <Label htmlFor="email">Email</Label>
-          <Input
-            id="email"
-            type="email"
-            placeholder="you@example.com"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-        </div>
-        <div className="space-y-2">
-          <Label htmlFor="password">Password</Label>
-          <Input
-            id="password"
-            type="password"
-            placeholder="Enter your password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-        </div>
-        <Button type="submit" className="w-full" disabled={loading}>
-          {loading ? "Signing in..." : "Sign in"}
-        </Button>
-      </form>
-
-      {signupEnabled && (
-        <p className="text-center text-sm text-muted-foreground">
-          Don&apos;t have an account?{" "}
-          <Link href="/auth/signup" className="text-primary underline-offset-4 hover:underline">
-            Sign up
-          </Link>
-        </p>
-      )}
-    </AuthShell>
-  );
+export default async function LoginPage() {
+  const signupEnabled = await getSignupEnabled();
+  return <LoginForm signupEnabled={signupEnabled} />;
 }
